// lib/openapi/generated/models/ProjectGeneralInformation.ts
export interface ProjectGeneralInformation {
    /**
     * Unique identifier
     * @type {number}
     * @memberof ProjectGeneralInformation
     */
    readonly id: number;
    /**
     * Project name / title
     * @type {string}
     * @memberof ProjectGeneralInformation
     */
    title?: string;
    /**
     * Name of the projects sponsor (Projektträger)
     * @type {string}
     * @memberof ProjectGeneralInformation
     */
    sponsor?: string;
    /**
     * Associated partner company
     * @type {Company}
     * @memberof ProjectGeneralInformation
     */
    partnerCompany?: Company;
    /**
     * Amount of contribution (Finanzierungssumme)
     * @type {MoneyAmount}
     * @memberof ProjectGeneralInformation
     */
    financingAmount?: MoneyAmount;
    /**
     * Start & Enddate of the projects runtime
     * @type {ProjectRuntime}
     * @memberof ProjectGeneralInformation
     */
    projectRuntime?: ProjectRuntime;
    /**
     * Name of the location / region, where the project takes place (Standort / Gebiet)
     * @type {string}
     * @memberof ProjectGeneralInformation
     */
    projectLocation?: string;
    /**
     * Name of the protected area that is included within the project region (Informationen bzgl. Schutzgebiet)
     * @type {string}
     * @memberof ProjectGeneralInformation
     */
    protectedArea?: string;
    /**
     * Coordinates of the project area
     * @type {Geocoordinates}
     * @memberof ProjectGeneralInformation
     */
    areaCoordinates?: Geocoordinates;
}